rewrite library to work much nicer
Yeah, this is my first experience with react, and I did a rewrite of the `snapblocks-react` component. This rewrite only imports snapblocks if it's actually running in the browser (or there is a `window` object). This is specifically because I was running into an error with serverside rendering in docusaurus.

Oh, and this rewrite actually works in docusaurus from the library. Better publish it now.